added executables

+++ src/light_sensor_node.cpp
@@ -4,7 +4,8 @@
 
 #include "../include/brickpi3_ros2/base_sensor_node.hpp"
 #include "../include/brickpi3_ros2/light_sensor_node.hpp"
-#include "../include/brickpi3_ros2/BrickPi3.hpp"
+
+#include "../libs/BrickPi3/BrickPi3.hpp"
 
 using namespace std::chrono;
 using namespace std;
@@ -12,7 +13,7 @@ using namespace std;
 LightSensorNode::LightSensorNode(): BaseSensorNode("light") {
     this->publisher = this->create_publisher<std_msgs::msg::Int16>("light", 10);
 
-    this->bp.set_sensor_type(this->port, SENSOR_TYPE_NXT_LIGHT_ON);
+    this->bp.set_sensor_type((uint8_t) this->port, (uint8_t) SENSOR_TYPE_NXT_LIGHT_ON);
 }
 
 void LightSensorNode::publish_value_callback() {
@@ -23,7 +24,7 @@ void LightSensorNode::publish_value_callback() {
 }
 
 int LightSensorNode::read_value_from_brickpi3() {
-    return this->bp.get_sensor(this->port, &this->sensor);
+    return this->bp.get_sensor((uint8_t) this->port, &this->sensor);
 }
 
 int main(int argc, char* argv[]) {
